requirements.txt
asgiref==3.5.2
Django==4.0.4
djangorestframework==3.14.0
gunicorn==20.1.0